moved sidebar component
