+++ src/components/Breadcrumbs/index.js
@@ -1,5 +1,9 @@
 import { Home, ChevronRight } from "@mui/icons-material";
+import MDButton from "components/MDButton";
 import "./style.scss";
+
+// Icons
+import LogoutIcon from "@mui/icons-material/Logout";
 
 const Breadcrumbs = ({ page }) => {
   return (
@@ -17,6 +21,16 @@ const Breadcrumbs = ({ page }) => {
           <ChevronRight />
           <span className="breadcrumbs__container__item__text">{page}</span>
         </div>
+      </div>
+      <div className="breadcrumbs__container">
+        <MDButton
+          buttonType="delete"
+          onClick={() => (window.location.href = "/")}
+          noIcon
+        >
+          <LogoutIcon style={{ marginRight: "5px" }} />
+          Logout
+        </MDButton>
       </div>
     </div>
   );
